+++ source/ftrack_api/_python_ntpath.py
@@ -1,3 +1,4 @@
+# pragma: no cover
 # Module 'ntpath' -- common operations on WinNT/Win95 pathnames
 """Common pathname manipulations, WindowsNT/95 version.